Settings: On-The-Go mode squashed [2/2]
On-The-Go turns on the camera of your phone and streams it to an overlay,
which is drawn over everything.
this makes you feel like your display is transparent, as you see the content
behind it.

its really useful when you are using your phone while walking, thats why
we called it "On-The-Go"

have fun!

Change-Id: I31c03d9e02c238e237681354b3991a0427c0c47b

f4c51f5 Settings: On-The-Go: Add quicktile [2/2]
877d505 Settings: On-The-Go mode [2/2]